[python] fix sending lines to ipython
In ipython >= 5, the autoindent feature is deprecated. Thus if we send
multiple lines directly to it, the indentation will go wild.
Notice that with ipyton >= 5, multiple line paste(via Ctrl-V or
Command-V) is well supported. This this commit try to utilize the `-p`
option of `tmux paste-buffer` which will use bracket paste mode.

Checkout here for more information: https://cirw.in/blog/bracketed-paste